Move IProfileAppService to account module
