+++ modules/account/src/Volo.Abp.Account.Application.Contracts/Volo/Abp/Account/ChangePasswordInput.cs
@@ -1,8 +1,9 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using Volo.Abp.Auditing;
+using Volo.Abp.Identity;
 using Volo.Abp.Validation;
 
-namespace Volo.Abp.Identity
+namespace Volo.Abp.Account
 {
     public class ChangePasswordInput
     {
