+++ modules/account/src/Volo.Abp.Account.Application.Contracts/Volo/Abp/Account/IProfileAppService.cs
@@ -1,7 +1,8 @@
 ﻿using System.Threading.Tasks;
 using Volo.Abp.Application.Services;
+using Volo.Abp.Identity;
 
-namespace Volo.Abp.Identity
+namespace Volo.Abp.Account
 {
     public interface IProfileAppService : IApplicationService
     {
