+++ modules/account/src/Volo.Abp.Account.Application/Volo/Abp/Account/AbpAccountApplicationModuleAutoMapperProfile.cs
@@ -1,0 +1,16 @@
+﻿using AutoMapper;
+using Volo.Abp.Identity;
+
+namespace Volo.Abp.Account
+{
+    public class AbpAccountApplicationModuleAutoMapperProfile : Profile
+    {
+        public AbpAccountApplicationModuleAutoMapperProfile()
+        {
+            CreateMap<IdentityUser, ProfileDto>()
+                .ForMember(dest => dest.HasPassword,
+                    op => op.MapFrom(src => src.PasswordHash != null))
+                .MapExtraProperties();
+        }
+    }
+}
