+++ modules/account/src/Volo.Abp.Account.HttpApi.Client/ClientProxies/ProfileClientProxy.Generated.cs
@@ -9,7 +9,7 @@ using Volo.Abp.Http.Client.ClientProxying;
 using Volo.Abp.Identity;
 
 // ReSharper disable once CheckNamespace
-namespace Volo.Abp.Identity.ClientProxies
+namespace Volo.Abp.Account.ClientProxies
 {
     [Dependency(ReplaceServices = true)]
     [ExposeServices(typeof(IProfileAppService), typeof(ProfileClientProxy))]
