+++ modules/account/src/Volo.Abp.Account.HttpApi.Client/ClientProxies/ProfileClientProxy.cs
@@ -1,6 +1,6 @@
 // This file is part of ProfileClientProxy, you can customize it here
 // ReSharper disable once CheckNamespace
-namespace Volo.Abp.Identity.ClientProxies
+namespace Volo.Abp.Account.ClientProxies
 {
     public partial class ProfileClientProxy
     {
