+++ modules/account/src/Volo.Abp.Account.Web/Pages/Account/Components/ProfileManagementGroup/Password/Default.js
@@ -23,7 +23,7 @@
                 return;
             }
 
-            volo.abp.identity.profile.changePassword(input).then(function (result) {
+            volo.abp.account.profile.changePassword(input).then(function (result) {
                 abp.message.success(l('PasswordChanged'));
             });
         });
