+++ modules/account/src/Volo.Abp.Account.Web/Pages/Account/Components/ProfileManagementGroup/PersonalInfo/Default.js
@@ -11,7 +11,7 @@
 
             var input = $('#PersonalSettingsForm').serializeFormToObject();
 
-            volo.abp.identity.profile.update(input).then(function (result) {
+            volo.abp.account.profile.update(input).then(function (result) {
                 abp.notify.success(l('PersonalSettingsSaved'));
             });
         });
