+++ modules/identity/src/Volo.Abp.Identity.Application/Volo/Abp/Identity/AbpIdentityApplicationModuleAutoMapperProfile.cs
@@ -1,5 +1,4 @@
 ﻿using AutoMapper;
-using Volo.Abp.AutoMapper;
 
 namespace Volo.Abp.Identity
 {
@@ -11,11 +10,6 @@ namespace Volo.Abp.Identity
                 .MapExtraProperties();
 
             CreateMap<IdentityRole, IdentityRoleDto>()
-                .MapExtraProperties();
-
-            CreateMap<IdentityUser, ProfileDto>()
-                .ForMember(dest => dest.HasPassword,
-                    op => op.MapFrom(src => src.PasswordHash != null))
                 .MapExtraProperties();
         }
     }
